requirements.txt
Django==3.2.11
django-storages[google]==1.12.3
django-environ==0.8.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
google-cloud-secret-manager==2.8.0
authlib
requests
pandas
sky-api-python-client
grpcio
cloud-sql-python-connector[pg8000]
pg8000
google-cloud-storage
djangorestframework
django-cors-headers
django-webpack-loader